pyproject.toml
[tool.poetry]
name = "gandynamic"
version = "0.1.0"
description = ""
authors = ["dek <dekonnection@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
PyYAML = "^6.0"
Cerberus = "^1.3.4"
deepmerge = "^1.0.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
